requirements.txt
flask
flask-cors
google-generativeai
openai
transformers
python-dotenv
requests
aiohttp 